entrypoint fix
Check for /etc/entrypoint.d/ before copying the data

+++ bin/docker-entrypoint.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 
 # prepare the etc dir
-cp -R /etc_entrypoint/* /etc/
+if [ ! -f "/etc/entrypoint.d" ];
+then
+		cp -R /etc_entrypoint/* /etc/
+fi
 
 # read entrypoint.d configuration
 for x in `find /etc/entrypoint.d -name '*.conf' | sort`;
